{"schema":"decodex/commit/1","summary":"Move scroll observation logging","authority":"manual"}
